Add access_token to cache
