+++ package.json
@@ -39,6 +39,7 @@
     "@react-email/components": "0.0.21",
     "@react-email/tailwind": "^0.0.18",
     "@t3-oss/env-nextjs": "^0.10.1",
+    "@tanstack/react-query": "^5.51.16",
     "@types/jest-axe": "^3.5.9",
     "axios": "^1.7.2",
     "class-variance-authority": "^0.7.0",
